pyproject.toml
[project]
name = "AtomForge"
version = "0.1.0"
description = "Materials science discovery platform."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "modal>=1.4.2",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "simpleeval==1.0.3",
    "numpy>=2.4.4",
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.44.0",
]

[project.scripts]
atomforge = "main:main"

[project.optional-dependencies]
physics = [
    # MLIP Core
    "torch==2.3.1",
    "mace-torch==0.3.15",
    "cuequivariance-torch",
    "cuequivariance",
    # Simulation & Data
    "ase==3.28.0",
    "pymatgen==2026.3.23",
    "mp-api",
    "emmet-core",
    "pandas>=2.2.0",
    "scipy",
]
dev = [
    "ruff>=0.15.11",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
